Allow instance processing only when a matching machine-id exists

+++ delete_all_cluster.sh
@@ -16,42 +16,45 @@ if [ -d "instances" ]
 then
   instance=$original_pwd/instances/`cat /var/lib/dbus/machine-id`
 
-  cd ${instance}
+  if [ -d "instance" ]
+  then
+    cd ${instance}
 
-  for cluster_name in * ;
-  do
-    if [ "$cluster_name" != "*" ] 
-    then	    
-      echo "Processing cluster $cluster_name at folder `pwd`"
+    for cluster_name in * ;
+    do
+      if [ "$cluster_name" != "*" ] 
+      then	    
+        echo "Processing cluster $cluster_name at folder `pwd`"
   
-      for terraform_repo_name in control_plane worker load_balancer ;
-      do
-        echo "  Check existenz at ${instance}/${cluster_name}/terraform/${terraform_repo_name}"
+        for terraform_repo_name in control_plane worker load_balancer ;
+        do
+          echo "  Check existenz at ${instance}/${cluster_name}/terraform/${terraform_repo_name}"
   
-        if [ -d ${instance}/${cluster_name}/terraform/${terraform_repo_name} ]
-        then
-          echo "    Terraform destroying $cluster_name -> $terraform_repo_name"
-  
-          cd ${instance}/${cluster_name}/terraform/${terraform_repo_name}
-  
-  	echo yes | terraform destroy
-  
+          if [ -d ${instance}/${cluster_name}/terraform/${terraform_repo_name} ]
+          then
+            echo "    Terraform destroying $cluster_name -> $terraform_repo_name"
+    
+            cd ${instance}/${cluster_name}/terraform/${terraform_repo_name}
+    
+      	  echo yes | terraform destroy
+    
           echo "    Deleting terraform folder"
   
-  	rm -rf ${instance}/${cluster_name}/terraform/${terraform_repo_name}
-        fi
-      done
+    	  rm -rf ${instance}/${cluster_name}/terraform/${terraform_repo_name}
+          fi
+        done
   
-      sudo virsh pool-destroy  ${cluster_name} 2> /dev/null
-      sudo virsh pool-delete   ${cluster_name} 2> /dev/null
-      sudo virsh pool-undefine ${cluster_name} 2> /dev/null
+        sudo virsh pool-destroy  ${cluster_name} 2> /dev/null
+        sudo virsh pool-delete   ${cluster_name} 2> /dev/null
+        sudo virsh pool-undefine ${cluster_name} 2> /dev/null
   
-      sudo rm -rf ${iso_node_path}/${cluster_name}
-      sudo rm -rf ${storage_path}/${cluster_name}
-  
-      rm -rf ${instance}/${cluster_name}
-    fi
-  done
+        sudo rm -rf ${iso_node_path}/${cluster_name}
+        sudo rm -rf ${storage_path}/${cluster_name}
+    
+        rm -rf ${instance}/${cluster_name}
+      fi
+    done
+  fi
 
   cd $original_pwd
 fi
